Refactor util function
