+++ packages/core/codeframe/package.json
@@ -17,8 +17,6 @@
   },
   "dependencies": {
     "chalk": "^2.4.2",
-    "emphasize": "^2.1.0",
-    "json-source-map": "^0.6.1",
-    "nullthrows": "^1.1.1"
+    "emphasize": "^2.1.0"
   }
 }
